best_stocks_project/requirements.txt
streamlit
fastapi
uvicorn
python-dotenv
yfinance
pandas
numpy
plotly
redis
requests
scipy
pypfopt
PyYAML
